notebook: add a second story cost calculator
Add a new Python module to the Jupyter notebook which takes the square
footage of a proposed second story addition and calculates an estimated
total cost of the addition. An optional parameter is added to this
module, which, when passed, will output a breakdown of the individual
costs for the second story addition.