notebook: add bedroom, bathroom cost calculators
Add a couple more Python modules to the Jupyter notebook which estimate
the costs of constructing a new bedroom and a new bathroom. Like the
second story calculator, both functions have an optional parameter
which, when passed, will output a breakdown of the individual costs for
the bedroom and bathroom additions.