notebook: add basic data visualization and model
Add the Housing .csv dataset to the repository for data visualization and modeling.

Add some code blocks to the notebook that load the .csv file and parse four parameters from the dataset based on the columns most related to the existing Python modules already in the notebook: area, number of bedrooms, bathrooms, and stories.

Add some basic data visualization and initialize a linear regression model based on these parameters. Create a visualization of the test data set and output performance metrics for it.